requirements.txt
google-generativeai
streamlit==1.27.2
watchdog